added news page and links to it on nav bar
