+++ src/App.js
@@ -3,6 +3,7 @@ import { Routes, Route, BrowserRouter } from 'react-router-dom';
 import ReadersPage from './Pages/ReadersPage.jsx'
 import Layout from './Components/Layout/Layout.jsx';
 import HomePage from './Pages/HomePage.jsx';
+import NewsPage from './Pages/NewsPage.jsx';
 import './App.css';
 
 function App() {
@@ -14,6 +15,7 @@ function App() {
           <Route path="/" element={<Layout />}>
             <Route index element={<HomePage />}></Route>
             <Route path="/Readers" element={<ReadersPage />}></Route>
+            <Route path="/News" element={<NewsPage />}></Route>
           </Route>
         </Routes>
       </BrowserRouter>
