+++ src/Components/PageElements/InfoCard/InfoCard.css
@@ -12,6 +12,7 @@
     justify-content: space-between;
     padding: 0% 10%;
     flex-wrap: wrap;
+    margin-bottom: 6%;
 }
 
 .infocard-flexbox {
