+++ src/Pages/NewsPage.jsx
@@ -1,0 +1,9 @@
+import { NewsCards } from "../Components/HomePage/NewsCard/NewsCards"
+
+export default function NewsPage() {
+    return (
+        <main className="newspage">
+            <NewsCards></NewsCards>
+        </main>
+    )
+}
